IGNITE-6932 move test to another suite
